Intergration of blocks previewer into the blocks preprocessor is completed
The blocks previewer is now fully integrated into the blocks preprocessor. The addition of the blocks.ts file for iteration over the book contents is now implemented, both previewer and preprocessor share the same code base allowing easy maintenance of both preprocessor and previewer. Will research into implementing a build pipeline where any new changes to the previewer are built and immediately integrated with the preprocessor.